moved UploadState to models, added tests for confirm_upload
